Extract the FileChangeWatcher into a workspace-level interface
This abstraction is consumed by a lot of the stuff in our VS layer
that needs to watch changes for files/references. We'll need a cross-
plat implementation of it, but the interface itself can be easily
abstracted so we can move more stuff around.
