+++ src/VisualStudio/Core/Def/ProjectSystem/RuleSets/VisualStudioRuleSetManager.cs
@@ -6,6 +6,7 @@
 
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Host;
+using Microsoft.CodeAnalysis.ProjectSystem;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Roslyn.Utilities;
 
@@ -14,14 +15,14 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.ProjectSystem
     internal sealed partial class VisualStudioRuleSetManager : IWorkspaceService
     {
         private readonly IThreadingContext _threadingContext;
-        private readonly FileChangeWatcher _fileChangeWatcher;
+        private readonly IFileChangeWatcher _fileChangeWatcher;
         private readonly IAsynchronousOperationListener _listener;
 
         private readonly ReferenceCountedDisposableCache<string, RuleSetFile> _ruleSetFileMap = new();
 
         public VisualStudioRuleSetManager(
             IThreadingContext threadingContext,
-            FileChangeWatcher fileChangeWatcher,
+            IFileChangeWatcher fileChangeWatcher,
             IAsynchronousOperationListener listener)
         {
             _threadingContext = threadingContext;
